refactor: extract run verification phase
